+++ src/anchor.hpp
@@ -29,7 +29,7 @@ private:
     */
     Eigen::VectorXi find_indeces()
     {
-        unsigned int n = m_data->rows();
+        size_t n = m_data->rows();
 
         m_origin_x = (m_data->col(0)).minCoeff()*(1 - Tolerances::anchor_tolerance);
         m_origin_y = (m_data->col(1)).minCoeff()*(1 - Tolerances::anchor_tolerance);
@@ -41,7 +41,7 @@ private:
 
         // fill a vector with the position of each point
         Eigen::VectorXi result(n);
-        for (unsigned int i=0; i<n; ++i)
+        for (size_t i=0; i<n; ++i)
         {
             cd::vector coordinates = m_data->row(i);
             result(i) = ceil((coordinates(0)-m_origin_x)/m_larghezza_cubo) + m_n_cubotti*floor((coordinates(1)-m_origin_y)/m_altezza_cubo);
@@ -62,23 +62,23 @@ public:
     */
     const cd::matrix find_anchorpoints()
     {
-        unsigned int n = m_data->rows();
+        size_t n = m_data->rows();
         Eigen::VectorXi indeces = find_indeces();
 
         // build a new vector without duplicates
-        std::vector<unsigned int> positions;
-        for (unsigned int i=0; i<n; ++i)
+        std::vector<size_t> positions;
+        for (size_t i=0; i<n; ++i)
         {
-            unsigned int pos = indeces(i);
+            size_t pos = indeces(i);
             if (std::find(positions.begin(), positions.end(), pos) == positions.end())
                 positions.push_back(pos);
         }
 
         // fill a new matrix with the coordinates of each anchorpoins
         cd::matrix anchorpos(positions.size(), m_data->cols());
-        for (unsigned int i=0; i<anchorpos.rows(); ++i)
+        for (size_t i=0; i<anchorpos.rows(); ++i)
         {
-            unsigned int I = positions[i];
+            size_t I = positions[i];
             anchorpos(i,0) = m_origin_x + (I - floor((I*(1 - Tolerances::anchor_tolerance))/m_n_cubotti)*m_n_cubotti)*m_larghezza_cubo - m_larghezza_cubo/2;
             anchorpos(i,1) = m_origin_y + ceil((I*(1 - Tolerances::anchor_tolerance))/m_n_cubotti)*m_altezza_cubo - m_altezza_cubo/2;
         }
